fix: store FSMEdge rule_id as int_32 to the cases with >32767 rules. (#652)
This pr changes use int32 to store edges instead of int16, which
supports cases with more than 32767 rules.

The benchmark code is as follows:

```python
"""Micro-benchmark for the FSM edge int16->int32 change.

Test point: `MainModel` JSON schema taken from
tests/python/test_grammar_matcher_json_schema.py (covers ints, floats, bools,
arrays, tuples, nested objects). Uses the non-gated gpt2 tokenizer so it runs
without a HF token / GPU.

Measures:
  * compilation time   : GrammarCompiler.compile_json_schema (fresh compiler/iter)
  * mask generation    : GrammarMatcher.fill_next_token_bitmask (CPU; this is the
                         FSM path that the edge struct change affects)
"""

import argparse
import statistics
import time
from typing import List, Tuple

from pydantic import BaseModel
from transformers import AutoTokenizer

import xgrammar as xgr


class MainModel(BaseModel):
    integer_field: int
    number_field: float
    boolean_field: bool
    any_array_field: list
    array_field: List[str]
    tuple_field: Tuple[str, int, List[str]]
    object_field: dict
    nested_object_field: dict


instance = MainModel(
    integer_field=42,
    number_field=3.14e5,
    boolean_field=True,
    any_array_field=[3.14, "foo", None, True],
    array_field=["foo", "bar"],
    tuple_field=("foo", 42, ["bar", "baz"]),
    object_field={"foo": 42, "bar": 43},
    nested_object_field={"foo": {"bar": 42}},
)
INSTANCE_STR = instance.model_dump_json(indent=2)


def bench(tokenizer_name: str, num_iters: int, num_warmup: int):
    tokenizer = AutoTokenizer.from_pretrained(tokenizer_name)
    tokenizer_info = xgr.TokenizerInfo.from_huggingface(tokenizer)
    vocab_size = tokenizer_info.vocab_size

    compile_times_ms: List[float] = []
    mask_times_us: List[float] = []
    num_mask_calls = len(INSTANCE_STR)

    for it in range(-num_warmup, num_iters):
        # --- compilation time: fresh compiler each iter to avoid cache hits ---
        compiler = xgr.GrammarCompiler(tokenizer_info, cache_enabled=False)
        t0 = time.perf_counter()
        compiled_grammar = compiler.compile_json_schema(MainModel, indent=2)
        t1 = time.perf_counter()

        # --- mask generation time: one fill_next_token_bitmask per char ---
        matcher = xgr.GrammarMatcher(compiled_grammar)
        bitmask = xgr.allocate_token_bitmask(1, vocab_size)
        t2 = time.perf_counter()
        for c in INSTANCE_STR:
            matcher.fill_next_token_bitmask(bitmask)
            assert matcher.accept_string(c), f"rejected char {c!r}"
        t3 = time.perf_counter()

        if it >= 0:
            compile_times_ms.append((t1 - t0) * 1e3)
            mask_times_us.append((t3 - t2) / num_mask_calls * 1e6)

    def summarize(xs):
        return statistics.mean(xs), statistics.median(xs), min(xs)

    c_mean, c_med, c_min = summarize(compile_times_ms)
    m_mean, m_med, m_min = summarize(mask_times_us)

    print(f"xgrammar serialize version : {xgr.get_serialization_version()}")
    print(f"tokenizer                  : {tokenizer_name} (vocab={vocab_size})")
    print(f"iters/warmup               : {num_iters}/{num_warmup}")
    print(f"mask calls per iter        : {num_mask_calls}")
    print("-" * 56)
    print(f"compilation time   (ms)    : mean={c_mean:.4f}  median={c_med:.4f}  min={c_min:.4f}")
    print(f"mask gen time  (us/token)  : mean={m_mean:.4f}  median={m_med:.4f}  min={m_min:.4f}")


if __name__ == "__main__":
    ap = argparse.ArgumentParser()
    ap.add_argument("--tokenizer", default="meta-llama/Llama-3.1-8B-Instruct")
    ap.add_argument("--num_iters", type=int, default=200)
    ap.add_argument("--num_warmup", type=int, default=20)
    args = ap.parse_args()
    bench(args.tokenizer, args.num_iters, args.num_warmup)
```

and the results are:

```text
int_32
xgrammar serialize version: v14
tokenizer                  : meta-llama/Llama-3.1-8B-Instruct (vocab=128256)
iters/warmup               : 200/20
mask calls per iter        : 392
--------------------------------------------------------
compilation time   (ms)    : mean=16.8092  median=16.9765  min=13.1094
mask gen time  (us/token)  : mean=16.6398  median=17.0990  min=9.3919

int_16
xgrammar serialize version: v13
tokenizer                  : meta-llama/Llama-3.1-8B-Instruct (vocab=128256)
iters/warmup               : 200/20
mask calls per iter        : 392
--------------------------------------------------------
compilation time   (ms)    : mean=16.8140  median=16.7889  min=13.3949
mask gen time  (us/token)  : mean=15.2320  median=16.3241  min=9.2153
```

The result demonstrates that it will not hurt the efficiency.

---------

+++ cpp/grammar_functor.cc
@@ -1476,7 +1476,7 @@ std::optional<FSMWithStartEnd> GrammarFSMBuilderImpl::TokenTagDispatch(
     auto [token_id, rule_id] = ttd.trigger_rule_pairs[i];
     fsm.AddTokenEdge(start, dispatch_state, {token_id});
     int target = loop ? start : end_state;
-    fsm.AddRuleEdge(dispatch_state, target, static_cast<int16_t>(rule_id));
+    fsm.AddRuleEdge(dispatch_state, target, static_cast<int32_t>(rule_id));
   }
   fsm.AddExcludeTokenEdge(start, start, self_loop_exclude);
   return FSMWithStartEnd(fsm, start, ends);
